Move the guts of instance_exec out into static methods
